运维功能完善

+++ tetris-project/omms/cloud/tetris-omms/src/main/java/com/sumavision/tetris/omms/software/service/type/ServiceTypeService.java
@@ -5,7 +5,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
-import org.hibernate.validator.internal.xml.GroupsType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,6 +15,7 @@ import com.sumavision.tetris.omms.software.service.installation.InstallationPack
 import com.sumavision.tetris.omms.software.service.installation.InstallationPackagePO;
 import com.sumavision.tetris.omms.software.service.installation.PropertiesDAO;
 import com.sumavision.tetris.omms.software.service.installation.PropertiesPO;
+import com.sumavision.tetris.omms.software.service.type.exception.DeploymentNotNullException;
 import com.sumavision.tetris.omms.software.service.type.exception.NoChoiceServerTypeException;
 import com.sumavision.tetris.omms.software.service.type.exception.NoGroupTyprOrNameException;
 import com.sumavision.tetris.omms.software.service.type.exception.NoServiceTypesToAddException;
@@ -124,34 +124,38 @@ public class ServiceTypeService {
 	}
 	
 	/**
-	 * 删除服务类型<br/>
+	 * 删除服务类型(已部署服务情况下不可删除)<br/>
 	 * <b>作者:</b>lqxuhv<br/>
 	 * <b>版本：</b>1.0<br/>
 	 * <b>日期：</b>2020年9月4日 上午11:13:36
 	 * @param id 服务类型id
 	 */
 	public void delete(Long id) throws Exception{
-		ServiceTypePO serviceTypePO = serviceTypeDao.findOne(id);
-		List<InstallationPackagePO> installationPackagePOs = installationPackageDAO.findByServiceTypeId(id);
-		for (InstallationPackagePO installationPackagePO : installationPackagePOs) {
-			List<PropertiesPO> propertiesPOs  = propertiesDAO.findByInstallationPackageId(installationPackagePO.getId());
-			for (PropertiesPO propertiesPO : propertiesPOs) {
-				if (propertiesPO != null) {
-					propertiesDAO.delete(propertiesPO);
+		List<ServiceDeploymentPO> serviceDeploymentPOs = serviceDeploymentDAO.findByServiceTypeId(id);
+		if (serviceDeploymentPOs !=null && !serviceDeploymentPOs.isEmpty()) {
+			throw new DeploymentNotNullException();
+		} else {
+			ServiceTypePO serviceTypePO = serviceTypeDao.findOne(id);
+			List<InstallationPackagePO> installationPackagePOs = installationPackageDAO.findByServiceTypeId(id);
+			for (InstallationPackagePO installationPackagePO : installationPackagePOs) {
+				List<PropertiesPO> propertiesPOs  = propertiesDAO.findByInstallationPackageId(installationPackagePO.getId());
+				for (PropertiesPO propertiesPO : propertiesPOs) {
+					if (propertiesPO != null) {
+						propertiesDAO.delete(propertiesPO);
+					}
+				}
+				if(installationPackagePO != null){
+					installationPackageDAO.delete(installationPackagePO);
 				}
 			}
-			if(installationPackagePO != null){
-				installationPackageDAO.delete(installationPackagePO);
+			for (ServiceDeploymentPO serviceDeploymentPO : serviceDeploymentPOs) {
+				if (serviceDeploymentPO != null) {
+					serviceDeploymentDAO.delete(serviceDeploymentPO);
+				}
 			}
-		}
-		List<ServiceDeploymentPO> serviceDeploymentPOs = serviceDeploymentDAO.findByServiceTypeId(id);
-		for (ServiceDeploymentPO serviceDeploymentPO : serviceDeploymentPOs) {
-			if (serviceDeploymentPO != null) {
-				serviceDeploymentDAO.delete(serviceDeploymentPO);
+			if(serviceTypePO != null){
+				serviceTypeDao.delete(serviceTypePO);
 			}
-		}
-		if(serviceTypePO != null){
-			serviceTypeDao.delete(serviceTypePO);
 		}
 	}
